docs: Make basic example more basic
Move existing "basic" example into new directory.
